+++ Year 1/Week02/3prog.c
