+++ Year 1/Week03/6prog.c
