+++ Year 1/Week07/2prog.c
